[Chore #162871744] Reorganizing app folder and file structure to prevent git merge conflicts
