--- .travis.yml
@@ -1,20 +1,0 @@
-env:
-  global:
-  - CC_TEST_REPORTER_ID=5df884ebcd2dbbd6d738e6139511816c3e81ae29a99e25d6dac1c4dc190dc178
-language: node_js
-node_js:
-- stable
-cache:
-  directories:
-  - node_modules
-before_script:
-- npm install
-- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-- chmod +x ./cc-test-reporter
-- "./cc-test-reporter before-build"
-script:
-- npm test
-- npm run coveralls
-- npm run codeClimate
-after_script:
-- "./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT"
