--- configs/express.js
@@ -1,14 +1,0 @@
-import express from 'express';
-import bodyParser from 'body-parser';
-import routes from '../routes/diaryRoute';
-
-const app = express();
-
-app.use(bodyParser.urlencoded({
-  extended: true,
-}));
-app.use(bodyParser.json());
-
-routes(app);
-
-export default app;
