--- helpers/diaryHelper.js
@@ -1,3 +1,0 @@
-class Helper {}
-
-export default Helper;
