+++ models/diaries.js
@@ -1,0 +1,7 @@
+const diaries = [{
+  id: 0,
+  titile: 'Title',
+  description: 'Description',
+}];
+
+export default diaries;
