--- models/diaryModel.js
@@ -1,6 +1,0 @@
-const diaryList = [{
-  id: 0,
-  title: 'breakfast',
-  description: 'alone',
-}];
-export default diaryList;
